Refactor workExperience data move to data folder
